requirements.txt
langchain==0.1.0
openai==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
streamlit==1.44.0
pillow==10.2.0
python-dotenv==1.0.0
tiktoken==0.5.1
boto3==1.34.0
nltk==3.8.1
python-docx==0.8.11
tqdm==4.66.1
pandas==2.0.0
torch==2.2.2
plotly==5.18.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai
tomli
